requirements.txt
transformers[sentencepiece]==4.53.0
sentencepiece==0.2.0
datasets==3.6.0
sacrebleu==2.5.1
rouge_score==0.1.2
py7zr==0.20.8
torch==2.7.1
pandas==2.2.3
numpy==2.2.6
ensure==1.0.4
python-box==7.3.2
PyYAML==6.0.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
mlflow==3.1.0
dagshub==0.5.10
flask==3.1.1
gunicorn==23.0.0